Move TransactionType model to shared/models directory
