+++ front-end/src/app/shared/models/transaction-types/PAC_RECEIPT.model.spec.ts
@@ -1,4 +1,4 @@
-import { TransactionType } from 'app/shared/interfaces/transaction-type.interface';
+import { TransactionType } from 'app/shared/models/transaction-types/transaction-type.model';
 import { SchATransaction, ScheduleATransactionTypes } from '../scha-transaction.model';
 import { PAC_RECEIPT } from './PAC_RECEIPT.model';
