+++ front-end/src/app/shared/models/transaction-types/TRANSFER.model.spec.ts
@@ -1,4 +1,4 @@
-import { TransactionType } from 'app/shared/interfaces/transaction-type.interface';
+import { TransactionType } from 'app/shared/models/transaction-types/transaction-type.model';
 import { SchATransaction, ScheduleATransactionTypes } from '../scha-transaction.model';
 import { TRANSFER } from './TRANSFER.model';
