+++ front-end/src/app/shared/models/transaction-types/UNREGISTERED_RECEIPT_FROM_PERSON.model.spec.ts
@@ -1,6 +1,6 @@
 import { UNREGISTERED_RECEIPT_FROM_PERSON } from './UNREGISTERED_RECEIPT_FROM_PERSON.model';
 import { SchATransaction, ScheduleATransactionTypes } from '../scha-transaction.model';
-import { TransactionType } from 'app/shared/interfaces/transaction-type.interface';
+import { TransactionType } from 'app/shared/models/transaction-types/transaction-type.model';
 
 describe('UNREGISTERED_RECEIPT_FROM_PERSON', () => {
   let transactionType: UNREGISTERED_RECEIPT_FROM_PERSON;
@@ -18,8 +18,7 @@ describe('UNREGISTERED_RECEIPT_FROM_PERSON', () => {
   it('#factory() should return a SchATransaction', () => {
     const txn: SchATransaction = transactionType.getNewTransaction();
     expect(txn.form_type).toBe('SA11AI');
-    expect(txn.transaction_type_identifier).toBe(
-      ScheduleATransactionTypes.UNREGISTERED_RECEIPT_FROM_PERSON);
+    expect(txn.transaction_type_identifier).toBe(ScheduleATransactionTypes.UNREGISTERED_RECEIPT_FROM_PERSON);
   });
   it('#generatePurposeDescription() should not be defined', () => {
     expect((transactionType as TransactionType).generatePurposeDescription).toBe(undefined);
